requirements.txt
# Kalshi Arbitrage Bot - Dependencies
# Core
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# HTTP & WebSocket
aiohttp>=3.9.0
websockets>=12.0
requests>=2.31.0
httpx>=0.27.0

# Data Processing
numpy>=1.26.0
pandas>=2.1.0
scipy>=1.11.0

# Optimization (Frank-Wolfe, Integer Programming)
ortools>=9.8.0  # Google OR-Tools for IP solver (free alternative to Gurobi)
cvxpy>=1.4.0    # Convex optimization
# gurobipy>=11.0.0  # Uncomment if you have Gurobi license

# Database
sqlalchemy>=2.0.0
aiosqlite>=0.19.0
alembic>=1.13.0

# Cryptography for API Auth
cryptography>=42.0.0

# Async utilities
asyncio-throttle>=1.0.2
aiocache>=0.12.0

# Monitoring & Logging
structlog>=24.1.0
prometheus-client>=0.19.0

# CLI & Configuration
click>=8.1.0
rich>=13.7.0
typer>=0.9.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
httpx>=0.27.0  # For testing HTTP calls

# Optional: LLM for dependency detection
openai>=1.12.0
anthropic>=0.18.0

# Development
black>=24.1.0
ruff>=0.2.0
mypy>=1.8.0
